Advanced Usage

For most use cases, the default API @option is sufficient, however, there are some specific cases requires one to use Configurations's advanced API.

Alias

Field Name Alias

Sometimes, the struct field name is defined using a UTF-8 character such as Ω. It may cause incompatibility when converting to other format such as JSON/TOML or just a Dict for JSON-based REST API.

You can define an alias for the field in this case using the following syntax

@option struct MyType
    "alpha"
    α::Int
    "omega"
    Ω::Float64 = 1.0
end

Option Type Alias

When there are multiple possible choices for an option-typed field, e.g

@option struct Options
    options::Union{OptionA, OptionB}
end

where OptionA and OptionB are also option types, one can specific which option type is it by using an alias when defining OptionA and OptionB

@option "A" struct OptionA
    name::String
end

@option "B" struct OptionB
    age::Int
end

then you can create an Options from the following Julia Dict

Dict{String, Any}(
    "options" => Dict{String, Any}(
        "A" => Dict{String, Any}(
            "name"=>"Roger",
        )
    )
)

or by using the following TOML file,

[options.A]
name="Roger"

Custom Option Macro

In some cases, you may not want all the features we defined by default in Configurations, such as the printing, etc.

In this case, you can construct your own macro using the code generation passes defined in Configurations. The code generation passes API starts with codegen_.

Configurations uses an intermediate representation to represent user defined option types, which is the OptionDef struct:

Configurations.OptionDef — Type

OptionDef

Type to represent the option type definition.

Configurations.OptionDef — Method

OptionDef(ex, alias=nothing)

Create an OptionDef from given Julia Expr

Builtin Code Generator

Configurations.codegen_struct_def — Function

codegen_struct_def(x::OptionDef)

generate the struct (composite type) definition.

Configurations.codegen_kw_fn — Function

codegen_kw_fn(x::OptionDef)

Generate the keyword constructor function definition.

Example

julia> ex = :(struct Foo
       x::Int = 2
       end);

julia> x = OptionDef(ex)
struct Foo
  x::Int = 2
end

julia> codegen_kw_fn(x)
:(function Foo(; x = 2)
      Foo(x)
  end)

Configurations.codegen_field_default — Function

codegen_field_default(x::OptionDef)

Generate the default value reflection method field_default.

Configurations.codegen_field_alias — Function

codegen_field_alias(x::OptionDef)

Generate the field alias method field_alias.

Configurations.codegen_alias — Function

codegen_alias(x::OptionDef)

Generate type alias method alias.

Configurations.codegen_create — Function

codegen_create(x::OptionDef)

Generate option type creation method create.

Configurations.codegen_isequal — Function

codegen_isequal(x::OptionDef)

Generate Base.:(==) to overload comparison operator to compare_options for given option type.

Configurations.codegen_show_toml_mime — Function

codegen_show_toml_mime(x::OptionDef)

Forward Base.show(io::IO, ::MIME"application/toml", x::YourOptionType) to to_toml.