Advanced Usage

For most use cases, the default API @option is sufficient, however, there are some specific cases requires one to use Configurations's advanced API.

Alias

Option Type Alias

When there are multiple possible choices for an option-typed field, e.g

@option struct Options
    options::Union{OptionA, OptionB}
end

where OptionA and OptionB are also option types, one can specific which option type is it by using an alias when defining OptionA and OptionB

@option "A" struct OptionA
    name::String
end

@option "B" struct OptionB
    age::Int
end

then you can create an Options from the following Julia Dict

Dict{String, Any}(
    "options" => Dict{String, Any}(
        "A" => Dict{String, Any}(
            "name"=>"Roger",
        )
    )
)

or by using the following TOML file,

[options.A]
name="Roger"

Custom Option Macro

In some cases, you may not want all the features we defined by default in Configurations, such as the printing, etc.

In this case, you can construct your own macro using the code generation passes defined in Configurations. The code generation passes API starts with codegen_.

Configurations uses an intermediate representation defined by Expronicon to represent user defined option types, which is the JLKwStruct struct.

Builtin Code Generator

Missing docstring.

Missing docstring for codegen_ast. Check Documenter's build log for details.

Missing docstring.

Missing docstring for codegen_create. Check Documenter's build log for details.

Configurations.codegen_is_option — Function

codegen_is_option(x::JLKwStruct)

Generate the is_option method.

Configurations.codegen_convert — Function

codegen_convert(x::JLKwStruct)

Generate Base.convert from AbstractDict{String} to the given option type.

Missing docstring.

Missing docstring for codegen_field_default. Check Documenter's build log for details.

Configurations.codegen_type_alias — Function

codegen_type_alias(def::JLKwStruct)

Generate type alias method type_alias.

Configurations.codegen_isequal — Function

codegen_isequal(x::JLKwStruct)

Generate Base.:(==) to overload comparison operator to compare_options for given option type.

Missing docstring.

Missing docstring for codegen_show. Check Documenter's build log for details.