Reference

Configurations.Maybe — Type

maybe of type T or nothing

Configurations.PartialDefault — Type

PartialDefault{F}

Type for non-constant default value, it depends on the value of another field that has default value.

Configurations.codegen_convert — Method

codegen_convert(x::JLKwStruct)

Generate Base.convert from AbstractDict{String} to the given option type.

Configurations.codegen_is_option — Method

codegen_is_option(x::JLKwStruct)

Generate the is_option method.

Configurations.codegen_isequal — Method

codegen_isequal(x::JLKwStruct)

Generate Base.:(==) to overload comparison operator to compare_options for given option type.

Configurations.codegen_option_type — Method

codegen_option_type(def::JLKwStruct)

Generate the Configurations option type definition from a given JLKwStruct created by Expronicon.

Configurations.codegen_type_alias — Method

codegen_type_alias(def::JLKwStruct)

Generate type alias method type_alias.

Configurations.compare_options — Method

compare_options(a, b, xs...)::Bool

Compare option types check if they are the same.

Configurations.convert_to_option — Method

convert_to_option(::Type{OptionType}, ::Type{ValueType}, x) where {OptionType, ValueType}

Convert x to type ValueType for option type OptionType. This is similar to Base.convert, when creating an instance of the option type, but one can use this to avoid type piracy.

Configurations.create — Method

create(::Type{T}; kwargs...) where T

Create an instance of option type T from kwargs. Similar to the default keyword argument constructor, but one can use this to create custom keyword argument constructor with extra custom keywords.

Configurations.field_alias — Method

field_alias(::Type{T}, name::Symbol) where {T}

Return field name alias of given option types.

Configurations.field_aliases — Method

field_alias(::Type{T}) where T

Return all field name alias of given option types.

Configurations.field_default — Method

field_default(::Type{T}, name::Symbol)

Return the default value of field name of an option type T.

Configurations.field_defaults — Method

field_defaults(::Type)

Return default values of given option types.

Configurations.field_keywords — Method

field_keywords(::Type{T}) where T

Return all the option type field names given T, error if there are duplicated sub-fields.

Configurations.from_dict — Method

from_dict(::Type{T}, d::AbstractDict{String}; kw...) where T

Convert dictionary d to an option type T, the valud of valid fields of T in this dictionary d can be override by keyword arguments.

Example

julia> @option struct OptionA
           name::String = "Sam"
           age::Int = 25
       end

julia> d = Dict{String, Any}(
           "name" => "Roger",
           "age" => 10,
       );

julia> from_dict(OptionA, d; age=25)
OptionA(;
    name = "Roger",
    age = 25,
)

Configurations.from_dict_inner — Method

from_dict_inner(::Type{T}, d::AbstractDict{String}) where T

Internal method to convert a dictionary (subtype of AbstractDict) to type T, this method will not check if T is an option type via is_option, and will not validate if all the required fields are available in the dict object.

Configurations.from_field_kwargs — Method

from_field_kwargs(::Type{T}; kw...) where T

Convert keyword arguments to given option type T using the field keyword convention.

Configurations.from_kwargs! — Method

from_kwargs!(d::AbstractDict{String}, ::Type{T}, prefix::Maybe{Symbol} = nothing; kw...) where T

Internal method for inserting keyword arguments to given dictionary object d. It will overwrite existing keys in d if it is specified by keyword argument.

Configurations.from_kwargs — Method

from_kwargs(::Type{T}; kw...) where T

Convert keyword arguments to given option type T using the underscore convention.

Configurations.from_kwargs — Method

from_kwargs(convention!, ::Type{T}; kw...) where T

Convert keyword arguments to given option type T using convention!. See also from_dict.

Convention

from_underscore_kwargs!: use _ to disambiguate subfields of the same name, this is the default behaviour.

from_field_kwargs!: do not disambiguate subfields, errors if there are disambiguity

Configurations.from_toml — Method

from_toml(::Type{T}, filename::String; kw...) where T

Convert a given TOML file filename to an option type T. Valid fields can be override by keyword arguments. See also from_dict.

Configurations.from_toml_if_exists — Method

from_toml_if_exists(::Type{T}, filename::String; kw...) where T

Similar to from_toml but will create the option instance via from_kwargs(T;kw...) instead of error if the file does not exist.

Configurations.from_underscore_kwargs — Method

from_underscore_kwargs(::Type{T}; kw...) where T

Convert keyword arguments to given option type T using the underscore convention.

Configurations.is_option — Method

is_option(x)

Check if x is an option type or not.

Configurations.pick_union — Method

pick_union(::Type, x) -> type, value

Pick a type T and its corresponding value from a Union. For option types it should be a dictionary type. The value can be furthur converted to this type T via convert_to_option or Base.convert.

Configurations.to_dict — Method

to_dict(option; include_defaults=false, exclude_nothing=false) -> OrderedDict

Convert an object to an OrderedDict.

Kwargs

include_defaults: include the default value.

exclude_nothing: exclude fields that have value nothing, this supersedes include_defaults when they are both true.

Tips

to_dict does not export fields that are of the same values as the defaults. In most cases, this should be the default behaviour, and users should not use include_defaults, however, this can be overridden by changing include_defaults to true.

Configurations.to_dict — Method

to_dict(::Type{T}, x, option::ConvertOption) where T

Convert x when x is inside an option type T. option is a set of options to determine the conversion behaviour. this can be overloaded to change the behaviour of to_dict(x; kw...).

to_dict(::Type{T}, x) where T

One can also use the 2-arg version when x is not or does not contain an option type for convenience.

Example

The following is a builtin overload to handle list of options.

function Configurations.to_dict(::Type{T}, x::Vector, option::ConvertOption) where T
    if is_option(eltype(x))
        return map(p->to_dict(T, p, include_defaults), x)
    else
        return x
    end
end

The following overloads the 2-arg to_dict to convert all VersionNumber to a String for all kinds of option types.

Configurations.to_dict(::Type, x::VersionNumber) = string(x)

Configurations.to_toml — Method

to_toml([f::Function], io::IO, option; sorted=false, by=identity, kw...)

Convert an instance option of option type to TOML and write it to IO. See to_dict for other valid keyword options. See also TOML.print in the stdlib for the explaination of sorted, by and f.

Configurations.to_toml — Method

to_toml([f::Function], filename::String, option; sorted=false, by=identity, kw...)

Convert an instance option of option type to TOML and write it to filename. See also TOML.print.

Configurations.to_toml — Method

to_toml(x; sorted=false, by=identity, kw...)

Convert an instance x of option type to TOML and write it to String. See also TOML.print.

to_toml does not export fields that are of the same values as the defaults. This can be overridden by changing include_defaults to true.

Configurations.type_alias — Method

type_alias(::Type{OptionType}) -> String

Return the alias name of given OptionType.

Configurations.underscore_keywords — Method

underscore_keywords(::Type{T}) where T

Return keywords given T using the underscore convention.

Configurations.@option — Macro

@option [alias::String] <struct def>

Define an option struct type. This will auto-generate methods that parse a given Dict{String} object (the keys must be of type String) into an instance of the struct type you defined. One can use alias string to distinguish multiple possible option type for the same field.

Example

One can define option type via @option macro with or without an alias.

julia> "Option A"
       @option "option_a" struct OptionA
           name::String
           int::Int = 1
       end

julia> "Option B"
       @option "option_b" struct OptionB
           opt::OptionA = OptionA(;name = "Sam")
           float::Float64 = 0.3
       end

and convert a dict to an option type via from_dict.

julia> d = Dict{String, Any}(
           "opt" => Dict{String, Any}(
               "name" => "Roger",
               "int" => 2,
           ),
           "float" => 0.33
       );

julia> option = from_dict(OptionB, d)
OptionB(;
    opt = OptionA(;
        name = "Roger",
        int = 2,
    ),
    float = 0.33,
)

when there are multiple possible option type for one field, one can use the alias to distinguish them

julia> @option struct OptionD
           opt::Union{OptionA, OptionB}
       end

julia> d1 = Dict{String, Any}(
               "opt" => Dict{String, Any}(
                   "option_b" => d
               )
           );

julia> from_dict(OptionD, d1)
OptionD(;
    opt = OptionB(;
        opt = OptionA(;
            name = "Roger",
            int = 2,
        ),
        float = 0.33,
    ),
)