Reference

Configurations.Maybe — Type

maybe of type T or nothing

Configurations.no_default — Constant

const for non default fields

Configurations.Field — Type

Field

Type to represent a field definition in option type.

Configurations.OptionDef — Type

OptionDef(ex, alias=nothing)

Create an OptionDef from given Julia Expr

Configurations.alias — Method

alias(::Type{OptionType}) -> String

Return the alias name of given OptionType.

Configurations.codegen — Method

generate Julia AST from OptionDef.

Configurations.codegen_alias — Method

codegen_alias(x::OptionDef)

Generate type alias method alias.

Configurations.codegen_convert — Method

codegen_convert(x::OptionDef)

Generate Base.convert from AbstractDict{String} to the given option type.

Configurations.codegen_create — Method

codegen_create(x::OptionDef)

Generate option type creation method create.

Configurations.codegen_field_alias — Method

codegen_field_alias(x::OptionDef)

Generate the field alias method field_alias.

Configurations.codegen_field_default — Method

codegen_field_default(x::OptionDef)

Generate the default value reflection method field_default.

Configurations.codegen_is_option — Method

codegen_is_option(x::OptionDef)

Generate the is_option method.

Configurations.codegen_isequal — Method

codegen_isequal(x::OptionDef)

Generate Base.:(==) to overload comparison operator to compare_options for given option type.

Configurations.codegen_kw_fn — Method

codegen_kw_fn(x::OptionDef)

Generate the keyword constructor function definition.

Example

julia> ex = :(struct Foo
       x::Int = 2
       end);

julia> x = OptionDef(ex)
struct Foo
  x::Int = 2
end

julia> codegen_kw_fn(x)
:(function Foo(; x = 2)
      Foo(x)
  end)

Configurations.codegen_show_text — Method

codegen_show_text(x::OptionDef)

Generate Base.show overloading for given type for the default printing syntax.

Configurations.codegen_show_toml_mime — Method

codegen_show_toml_mime(x::OptionDef)

Forward Base.show(io::IO, ::MIME"application/toml", x::YourOptionType) to to_toml.

Configurations.codegen_struct_def — Method

codegen_struct_def(x::OptionDef)

generate the struct (composite type) definition.

Configurations.compare_options — Method

compare_options(a, b, xs...)::Bool

Compare option types check if they are the same.

Configurations.create — Method

create(::Type{T}; kwargs...) where T

Create an instance of option type T from kwargs. Similar to the default keyword argument constructor, but one can use this to create custom keyword argument constructor with extra custom keywords.

Configurations.dictionalize — Method

dictionalize(x; include_defaults=false)

Convert x to an OrderedDict.

Configurations.field_alias — Method

field_alias(::Type{T}, name::Symbol) where {T}

Return field name alias of given option types.

Configurations.field_aliases — Method

field_alias(::Type{T}) where T

Return all field name alias of given option types.

Configurations.field_default — Method

field_default(::Type{T}, name::Symbol)

Return the default value of field name of an option type T.

Configurations.field_defaults — Method

field_defaults(::Type)

Return default values of given option types.

Configurations.from_dict — Method

from_dict(::Type{T}, d::AbstractDict{String}; kw...) where T

Convert dictionary d to an option type T, the valud of valid fields of T in this dictionary d can be override by keyword arguments.

Example

julia> @option struct OptionA
           name::String = "Sam"
           age::Int = 25
       end

julia> d = Dict{String, Any}(
           "name" => "Roger",
           "age" => 10,
       );

julia> from_dict(OptionA, d; age=25)
OptionA(;
    name = "Roger",
    age = 25,
)

Configurations.from_dict_inner — Method

from_dict_inner(::Type{T}, d::AbstractDict{String}) where T

Internal method to convert a dictionary (subtype of AbstractDict) to type T, this method will not check if T is an option type via is_option, and will not validate if all the required fields are available in the dict object.

Configurations.from_kwargs! — Method

from_kwargs!(d::AbstractDict{String}, ::Type{T}, prefix::Maybe{Symbol} = nothing; kw...) where T

Internal method for inserting keyword arguments to given dictionary object d. It will overwrite existing keys in d if it is specified by keyword argument.

Configurations.from_kwargs — Method

from_kwargs(::Type{T}; kw...) where T

Convert keyword arguments to given option type T. See also from_dict.

Configurations.from_toml — Method

from_toml(::Type{T}, filename::String; kw...) where T

Convert a given TOML file filename to an option type T. Valid fields can be override by keyword arguments. See also from_dict.

Configurations.is_option — Method

is_option(x)

Check if x is an option type or not.

Configurations.keywords — Method

keywords(::Type{T}) where T -> Vector{Symbol}

Get all the keywords of type T.

Configurations.option_convert — Method

option_convert(::Type{OptionType}, ::Type{ValueType}, x) where {OptionType, ValueType}

Convert x to type ValueType for option type OptionType. This is similar to Base.convert, when creating an instance of the option type, but one can use this to avoid type piracy.

Configurations.pick_union — Method

pick_union(::Type, x) -> type, value

Pick a type T and its corresponding value from a Union. For option types it should be a dictionary type. The value can be furthur converted to this type T via option_convert or Base.convert.

Configurations.split_body — Method

split_body(ex::Expr) -> fields::Vector{Field}, misc::Vector{Any}

Split the fields of option type declaration and misc (such as inner constructors etc.).

Configurations.split_name — Method

split_name(ex::Expr) -> name, typevars, supertype

Split the name, type parameters and supertype definition from struct declaration head.

Configurations.to_dict — Method

to_dict(option; include_defaults=false) -> OrderedDict

Convert an option to an OrderedDict.

Tips

to_dict does not export fields that are of the same values as the defaults. In most cases, this should be the default behaviour, and users should not use include_defaults, however, this can be overridden by changing include_defaults to true.

Configurations.to_toml — Method

to_toml([f::Function], io::IO, option; sorted=false, by=identity, include_defaults=false)

Convert an instance option of option type to TOML and write it to IO. See also TOML.print.

Configurations.to_toml — Method

to_toml([f::Function], filename::String, option; sorted=false, by=identity, include_defaults=false)

Convert an instance option of option type to TOML and write it to filename. See also TOML.print.

Configurations.to_toml — Method

to_toml(x; sorted=false, by=identity, include_defaults=false)

Convert an instance x of option type to TOML and write it to String. See also TOML.print.

to_toml does not export fields that are of the same values as the defaults. This can be overridden by changing include_defaults to true.

Configurations.toml_convert — Method

toml_convert(::Type, x)

A convenient function for converting common Julia types to TOML compatible types. One can overload the first argument to custom the behaviour for a specific option type.

Configurations.toml_convert — Method

toml_convert(::Type{T}) where T

Curried version of toml_convert.

Configurations.@option — Macro

@option [alias::String] <struct def>

Define an option struct type. This will auto-generate methods that parse a given Dict{String} object (the keys must be of type String) into an instance of the struct type you defined. One can use alias string to distinguish multiple possible option type for the same field.

Example

One can define option type via @option macro with or without an alias.

julia> "Option A"
       @option "option_a" struct OptionA
           name::String
           int::Int = 1
       end

julia> "Option B"
       @option "option_b" struct OptionB
           opt::OptionA = OptionA(;name = "Sam")
           float::Float64 = 0.3
       end

and convert a dict to an option type via from_dict.

julia> d = Dict{String, Any}(
           "opt" => Dict{String, Any}(
               "name" => "Roger",
               "int" => 2,
           ),
           "float" => 0.33
       );

julia> option = from_dict(OptionB, d)
OptionB(;
    opt = OptionA(;
        name = "Roger",
        int = 2,
    ),
    float = 0.33,
)

when there are multiple possible option type for one field, one can use the alias to distinguish them

julia> @option struct OptionD
           opt::Union{OptionA, OptionB}
       end

julia> d1 = Dict{String, Any}(
               "opt" => Dict{String, Any}(
                   "option_b" => d
               )
           );

julia> from_dict(OptionD, d1)
OptionD(;
    opt = OptionB(;
        opt = OptionA(;
            name = "Roger",
            int = 2,
        ),
        float = 0.33,
    ),
)